type conversion

+++ db/BookmarkTable.sql
@@ -1,7 +1,7 @@
 create table BookmarkTable(
-    bookmarkId int not null auto_increment,
-    userId int not null,
-    postId int not null,
+    bookmarkId BIGINT not null auto_increment,
+    userId BIGINT not null,
+    postId BIGINT not null,
     primary key(bookmarkId),
     foreign key(userId) references UserTable(userId),
     foreign key(postId) references postTable(postId)
